DI:     -Fixed a filemode related bug (i.e. SSBB, MP3)

+++ di/dip.h
@@ -107,7 +107,7 @@ typedef struct
 
 typedef struct
 {
-	u64 Offset;
+	u32 Offset;
 	u32 Size;
 	FIL File;
 } FileCache;
